Challenge 3: Ballon, booms and booleans
In this new challenge, there is a started project with a lot of errors that have to be fixed.

This initial commit includes the fix of spaceBar input in order to float when pressing the key. It wasn´t communicating with rigidBody because it was not initialized.

Also, we added the ForceMode.Impulse in void function in order to jump in a more accurate way