Fixed errors and added a BONUS part
The function to invoke was mispelled, everything else was ok.

Also we moved the fireworks position in order to display them in players place.

We tried to fix the offBound limits in Y axis by allowing jumping only when it is less than screen limit (15 Y) but if the user jumps a lot from the bottom, the ballon will disappear for a few seconds while gravity does its work